+++ src/app/auth/sign-up/page.tsx
@@ -2,11 +2,14 @@
 
 import { AuthLayout } from '@/components/auth/shared';
 import { SignUpWizard } from '@/components/auth/sign-up';
+import { Suspense } from 'react';
 
 export default function SignUpPage() {
 	return (
 		<AuthLayout>
-			<SignUpWizard />
+			<Suspense fallback={null}>
+				<SignUpWizard />
+			</Suspense>
 		</AuthLayout>
 	);
 }